third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-addIceCandidate-timing.https-expected.txt

This is a testharness.js-based test.
PASS addIceCandidate is not resolved first if 2x setLocalDescription operations are pending
FAIL addIceCandidate and setLocalDescription are resolved in the correct order, as defined by the operations chain specification assert_array_equals: property 1, expected "setLocalDescription 2" but got "addIceCandidate"
Harness: the test ran to completion.
